<food> fixed bug with long
git-svn-id: file:///var/pocketcampus/trunk@1456 fd477b11-563a-408f-9d6e-eeb9ca009d0c

+++ plugin/food/plugin.food.server/src/org/pocketcampus/plugin/food/server/db/FoodDB.java
@@ -247,7 +247,7 @@ public class FoodDB {
 			String restaurant = null;
 			double totalRating = 0;
 			int numberOfVotes = 0;
-			int mealId = 0;
+			long mealId = 0;
 
 			// Treat the answer from the database
 			while (rset.next()) {
@@ -256,7 +256,7 @@ public class FoodDB {
 				restaurant = rset.getString("Restaurant");
 				totalRating = rset.getFloat("TotalRating");
 				numberOfVotes = rset.getInt("NumberOfVotes");
-				mealId = rset.getInt("MealId");
+				mealId = rset.getLong("MealId");
 
 				// Create a new meal from the info we got in the database
 				Rating mealRating = new Rating(FoodUtils.totalRatingToRatingValue(
